# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/group/create": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupInfo",
                "parameters": [
                    {
                        "description": "GroupCreateReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GroupCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GroupCreateResp"
                        }
                    }
                }
            }
        },
        "/group/list": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupInfo",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.GroupResp"
                            }
                        }
                    }
                }
            }
        },
        "/group/search": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupInfo",
                "parameters": [
                    {
                        "description": "GroupSearchReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GroupSearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/group/usersAdd": {
            "post": {
                "tags": [
                    "user2group"
                ],
                "summary": "UsersAddToGroup",
                "parameters": [
                    {
                        "description": "UsersInGroupReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UsersInGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
        "/group/usersDelete": {
            "post": {
                "tags": [
                    "user2group"
                ],
                "summary": "UsersDeleteFromGroup",
                "parameters": [
                    {
                        "description": "UsersInGroupReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UsersInGroupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
        "/group/usersList": {
            "get": {
                "tags": [
                    "user2group"
                ],
                "summary": "UsersListByGroup",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "groupId",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UsersResp"
                        }
                    }
                }
            }
        },
        "/group/{groupId}": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupInfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "groupId",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GroupResp"
                        }
                    }
                }
            }
        },
        "/group/{groupId}/delete": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupInfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "groupId",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
        "/group/{groupId}/update": {
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "GroupInfo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "groupId",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GroupUpdateReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GroupUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
...
        "/user/groupsAdd": {
            "post": {
                "tags": [
                    "user2group"
                ],
                "summary": "GroupsAddToUser",
                "parameters": [
                    {
                        "description": "GroupsInUserReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GroupsInUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
        "/user/groupsDelete": {
            "post": {
                "tags": [
                    "user2group"
                ],
                "summary": "GroupsDeleteFromUser",
                "parameters": [
                    {
                        "description": "GroupsInUserReq",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GroupsInUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ModifyRowsResp"
                        }
                    }
                }
            }
        },
        "/user/groupsList": {
            "get": {
                "tags": [
                    "user2group"
                ],
                "summary": "GroupsListByUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GroupsResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "condition.QueryCondition": {
            "type": "object",
            "required": [
                "comparison",
                "fieldname",
                "operator",
                "value"
            ],
            "properties": {
                "comparison": {
                    "type": "string",
                    "enum": [
                        "eq",
                        "ne",
                        "gt",
                        "ge",
                        "lt",
                        "le",
                        "in",
                        "like"
                    ]
                },
                "fieldname": {
                    "type": "string"
                },
                "operator": {
                    "type": "string",
                    "enum": [
                        "and",
                        "or"
                    ]
                },
                "value": {}
            }
        },
        "types.GroupCreateReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "types.GroupCreateResp": {
            "type": "object",
            "properties": {
                "lastId": {
                    "type": "integer"
                }
            }
        },
        "types.GroupResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.GroupSearchReq": {
            "type": "object",
            "properties": {
                "query": {
                    "description": "pagination.Page",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/condition.QueryCondition"
                    }
                }
            }
        },
        "types.GroupUpdateReq": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.GroupsInUserReq": {
            "type": "object",
            "required": [
                "groupIds",
                "userId"
            ],
            "properties": {
                "groupIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "types.GroupsResp": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "types.ModifyRowsResp": {
            "type": "object",
            "properties": {
                "rowsAffected": {
                    "type": "integer"
                }
            }
        },
...
        "types.UsersInGroupReq": {
            "type": "object",
            "required": [
                "groupId",
                "userIds"
            ],
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "userIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "types.UsersResp": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}